Cloud_pipelines/Azure/pca-pipeline/pca-pipeline/requirements.txt
numpy
azure-storage-blob==1.5.0
